requirements.txt
celery==5.0.2
coverage==5.3
Django==3.1.3
django-celery-results==2.0.0
django-csp==3.7
django-environ==0.4.5
django-localflavor==3.0.1
django-extensions==3.0.9
django-filter==2.4.0
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg[validation]==1.20.0
Faker==4.17.1
gevent==20.9.0
gunicorn==20.0.4
ipdb==0.13.4
ipython==7.19.0
model-bakery==1.2.1
pip==20.2.4
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-sugar==0.9.4
pytest-xdist==2.1.0
PyYAML==5.3.1
requests==2.25.0
responses==0.12.1
whitenoise==5.2.0
